fix(rt): TYCHO_BLOCK's guard was undefined on Windows -- long is 32-bit there
runtime/tycho_rt.c parsed TYCHO_BLOCK by hand into a `long`, capped with
`v < (1L << 40)`. `long` is 64-bit on Linux and 32-bit on Windows (LLP64),
so in every shipped .exe the shift is UB and the accumulator cannot reach
the cap it is checked against. Both widen to `long long`; widening only the
shift would leave the accumulator overflowing before the cap is consulted.

Found by `make mingw-warn`, which is the only lane that reads the cross
compiler's output, on its first Linux run:

  emitted/tychoc.c:607:95: warning: left shift count >= width of type
  [-Wshift-count-overflow]      (and tycho-debug, tychofmt, tycho-lsp)

Pre-existing since 7d2c9d39, identical on both parents of the merge in
17b88404 -- not introduced by it.

Modelled the guard at both widths (a 32-bit `long` truncates 65536 to 655,
1099511627776 to 1099) and confirmed the fixed form: 65536 exact, a 20-digit
input clamped at 9999999999999, "12abc" stops at 12, empty/0 give 0.

  make mingw-warn: ok (4 programs cross-compiled, 0 warning lines,
                       matching scripts/mingw.warn)
  make test:  passed 1037  failed 0
  make conc:  passed 48    failed 0
  make ilp32: passed 1037  failed 0, all green
  fuzz/run.py 200:        ok=200 skip=0 timeout=0 FAIL=0
  fuzz/run_reject.py 200: accepted=17 rejected=183 FAIL=0
  fuzz/run_leak.py 150:   ok=150 skip=0 FAIL=0

+++ runtime/tycho_rt.c
@@ -604,7 +604,10 @@ __attribute__((constructor)) static void stats_init(void) {
     /* TYCHO_BLOCK: override the default block size (bytes); 0/absent = default.
      * Digits by hand: glibc redirects strtol to __isoc23_strtol@GLIBC_2.38. */
     const char *bs = getenv("TYCHO_BLOCK");
-    if (bs && *bs) { long v = 0; const char *q = bs; while (*q >= '0' && *q <= '9' && v < (1L << 40)) v = v * 10 + (*q++ - '0'); if (q != bs && v > 0) g_block_override = (size_t)v; }
+    /* `long long`, not `long`: long is 32-bit on Windows (LLP64), where both
+     * the accumulator and `1L << 40` are undefined -- the cap could not be
+     * reached and the shift is UB. Caught by `make mingw-warn`. */
+    if (bs && *bs) { long long v = 0; const char *q = bs; while (*q >= '0' && *q <= '9' && v < (1LL << 40)) v = v * 10 + (*q++ - '0'); if (q != bs && v > 0) g_block_override = (size_t)v; }
 }
 
 static __thread HBlock *g_block_pool = NULL;
